# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resume AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Document {
  id: number;
  name: string;
  /** resume, cover_letter, portfolio, other */
  type: string;
  content: string;
  createdAt: string;
}

export interface DocumentInput {
  name: string;
  /** resume, cover_letter, portfolio, other */
  type: string;
  content: string;
}

export interface Job {
  id: number;
  url: string;
  /** @nullable */
  title: string | null;
  /** @nullable */
  company: string | null;
  /** @nullable */
  location: string | null;
  /** @nullable */
  description: string | null;
  /** pending, scraped, failed */
  status: string;
  createdAt: string;
}

export interface JobInput {
  url: string;
}

export interface Application {
  id: number;
  jobId: number;
  /** @nullable */
  jobTitle: string | null;
  /** @nullable */
  jobCompany: string | null;
  /** @nullable */
  resume: string | null;
  /** @nullable */
  coverLetter: string | null;
  /** generating, done, failed */
  status: string;
  createdAt: string;
}

export interface ApplicationInput {
  jobId: number;
}

export interface Stats {
  documentCount: number;
  jobCount: number;
  applicationCount: number;
  pendingJobs: number;
  completedApplications: number;
}