# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Resume AI API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: documents
    description: Uploaded resume/cover letter/portfolio documents
  - name: jobs
    description: Job listings
  - name: applications
    description: Generated resumes and cover letters
  - name: stats
    description: Dashboard statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all uploaded documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a document (text content upload)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all job listings
      responses:
        "200":
          description: List of job listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Add a job listing by URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created job listing (scraping may be async)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: Invalid URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List all generated applications
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Generate a tailored resume and cover letter for a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Generated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get a generated application with full content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application with full content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteApplication
      tags: [applications]
      summary: Delete an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Document:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          description: "resume, cover_letter, portfolio, other"
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - content
        - createdAt

    DocumentInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          description: "resume, cover_letter, portfolio, other"
        content:
          type: string
      required:
        - name
        - type
        - content

    Job:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        title:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        status:
          type: string
          description: "pending, scraped, failed"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - url
        - title
        - company
        - location
        - description
        - status
        - createdAt

    JobInput:
      type: object
      properties:
        url:
          type: string
      required:
        - url

    Application:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        jobTitle:
          type: ["string", "null"]
        jobCompany:
          type: ["string", "null"]
        resume:
          type: ["string", "null"]
        coverLetter:
          type: ["string", "null"]
        status:
          type: string
          description: "generating, done, failed"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - jobId
        - jobTitle
        - jobCompany
        - resume
        - coverLetter
        - status
        - createdAt

    ApplicationInput:
      type: object
      properties:
        jobId:
          type: integer
      required:
        - jobId

    Stats:
      type: object
      properties:
        documentCount:
          type: integer
        jobCount:
          type: integer
        applicationCount:
          type: integer
        pendingJobs:
          type: integer
        completedApplications:
          type: integer
      required:
        - documentCount
        - jobCount
        - applicationCount
        - pendingJobs
        - completedApplications