# lib/api-zod/src/generated/types/application.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resume AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Application {
  id: number;
  jobId: number;
  /** @nullable */
  jobTitle: string | null;
  /** @nullable */
  jobCompany: string | null;
  /** @nullable */
  resume: string | null;
  /** @nullable */
  coverLetter: string | null;
  /** generating, done, failed */
  status: string;
  createdAt: Date;
}