# lib/api-zod/src/generated/types/documentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resume AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentInput {
  name: string;
  /** resume, cover_letter, portfolio, other */
  type: string;
  content: string;
}