# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resume AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Job {
  id: number;
  url: string;
  /** @nullable */
  title: string | null;
  /** @nullable */
  company: string | null;
  /** @nullable */
  location: string | null;
  /** @nullable */
  description: string | null;
  /** pending, scraped, failed */
  status: string;
  createdAt: Date;
}