# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resume AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  documentCount: number;
  jobCount: number;
  applicationCount: number;
  pendingJobs: number;
  completedApplications: number;
}